+++ tests/RulesTest.php
@@ -114,7 +114,7 @@ class RulesTest extends TestCase
      * @covers \Pdp\PublicSuffix::setSection
      * @covers \Pdp\PublicSuffix::isKnown
      * @covers \Pdp\Domain::isKnown
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testNullWillReturnNullDomain()
     {
@@ -125,7 +125,7 @@ class RulesTest extends TestCase
 
     /**
      * @covers ::resolve
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testThrowsTypeErrorOnWrongInput()
     {
@@ -152,7 +152,7 @@ class RulesTest extends TestCase
      * @covers \Pdp\PublicSuffix::setSection
      * @covers \Pdp\PublicSuffix::isKnown
      * @covers \Pdp\Domain::isKnown
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testIsSuffixValidFalse()
     {
@@ -174,7 +174,7 @@ class RulesTest extends TestCase
      * @covers \Pdp\Domain::isKnown
      * @covers \Pdp\Domain::isICANN
      * @covers \Pdp\Domain::isPrivate
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testIsSuffixValidTrue()
     {
@@ -198,7 +198,7 @@ class RulesTest extends TestCase
      * @covers \Pdp\Domain::isKnown
      * @covers \Pdp\Domain::isICANN
      * @covers \Pdp\Domain::isPrivate
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testIsSuffixValidFalseWithPunycoded()
     {
@@ -222,7 +222,7 @@ class RulesTest extends TestCase
      * @covers \Pdp\Domain::isKnown
      * @covers \Pdp\Domain::isICANN
      * @covers \Pdp\Domain::isPrivate
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testSubDomainIsNull()
     {
@@ -235,7 +235,7 @@ class RulesTest extends TestCase
     /**
      * @covers ::resolve
      * @covers ::validateSection
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testWithExceptionName()
     {
@@ -249,7 +249,7 @@ class RulesTest extends TestCase
      * @covers ::findPublicSuffix
      * @covers ::findPublicSuffixFromSection
      * @covers \Pdp\PublicSuffix::setSection
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testWithPrivateDomain()
     {
@@ -280,7 +280,7 @@ class RulesTest extends TestCase
      * @covers ::findPublicSuffix
      * @covers ::findPublicSuffixFromSection
      * @covers \Pdp\PublicSuffix::setSection
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testWithPrivateDomainInvalid()
     {
@@ -298,7 +298,7 @@ class RulesTest extends TestCase
      * @covers ::findPublicSuffix
      * @covers ::findPublicSuffixFromSection
      * @covers \Pdp\PublicSuffix::setSection
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testWithPrivateDomainValid()
     {
@@ -316,7 +316,7 @@ class RulesTest extends TestCase
      * @covers ::findPublicSuffix
      * @covers ::findPublicSuffixFromSection
      * @covers \Pdp\PublicSuffix::setSection
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testWithICANNDomainInvalid()
     {
@@ -334,7 +334,7 @@ class RulesTest extends TestCase
      * @covers ::findPublicSuffix
      * @covers ::findPublicSuffixFromSection
      * @covers \Pdp\PublicSuffix::setSection
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testWithDomainObject()
     {
@@ -350,7 +350,7 @@ class RulesTest extends TestCase
 
     /**
      * @covers ::getPublicSuffix
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testWithDomainInterfaceObject()
     {
@@ -378,7 +378,7 @@ class RulesTest extends TestCase
 
     /**
      * @covers ::resolve
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      * @covers \Pdp\Domain::setPublicSuffix
      * @covers \Pdp\Domain::getPublicSuffix
      * @dataProvider parseDataProvider
@@ -395,7 +395,7 @@ class RulesTest extends TestCase
 
     /**
      * @covers ::resolve
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      * @covers \Pdp\Domain::withPublicSuffix
      * @covers \Pdp\Domain::getContent
      * @dataProvider parseDataProvider
@@ -440,7 +440,7 @@ class RulesTest extends TestCase
      * @covers ::getPublicSuffix
      * @covers ::validateSection
      * @covers \Pdp\Domain::isResolvable
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      * @dataProvider invalidParseProvider
      *
      * @param mixed $domain
@@ -465,7 +465,7 @@ class RulesTest extends TestCase
      * @covers ::getPublicSuffix
      * @covers ::validateSection
      * @covers \Pdp\Domain::isResolvable
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      * @dataProvider invalidDomainParseProvider
      *
      * @param mixed $domain
@@ -497,7 +497,7 @@ class RulesTest extends TestCase
      * @covers ::getPublicSuffix
      * @covers ::validateSection
      * @covers \Pdp\Domain::isResolvable
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      * @dataProvider validPublicSectionProvider
      *
      * @param string|null $domain
@@ -556,7 +556,7 @@ class RulesTest extends TestCase
      * @covers ::findPublicSuffixFromSection
      * @covers \Pdp\Domain::withPublicSuffix
      * @covers \Pdp\Domain::getRegistrableDomain
-     * @covers \Pdp\IDNAConverterTrait::setLabels
+     * @covers \Pdp\IDNAConverterTrait::parse
      */
     public function testPublicSuffixSpec()
     {